+++ src/lib/components/properties-menu/models/property-tree-config.model.ts
@@ -1,18 +1,26 @@
 import { PropertyTreeNode } from './property-tree-node.model';
 
 export interface PropertyTreeConfigParameters {
+    acceptsRootDrop?: boolean;
     addBlockLabel?: string;
     nodes?: PropertyTreeNode[];
     showEmptyStateAddBlock?: boolean;
 }
 
 export class PropertyTreeConfig {
+    acceptsRootDrop: boolean;
     nodes: PropertyTreeNode[];
     showEmptyStateAddBlock: boolean;
 
     addBlockLabel?: string;
 
-    constructor({ addBlockLabel, nodes = [], showEmptyStateAddBlock = false }: PropertyTreeConfigParameters) {
+    constructor({
+        acceptsRootDrop = false,
+        addBlockLabel,
+        nodes = [],
+        showEmptyStateAddBlock = false
+    }: PropertyTreeConfigParameters) {
+        this.acceptsRootDrop = acceptsRootDrop;
         this.addBlockLabel = addBlockLabel;
         this.nodes = nodes;
         this.showEmptyStateAddBlock = showEmptyStateAddBlock;